Update promise.js
Hi Juan,

I just found out (using FF 30.0 on Mac), that native Promise DOES log any JS-error which occur inside a Promise which has no error-callbacks (with a delay of several seconds). Personaly, I think this is huge. I don't know if it's just FF or if it's a Promise A+ specification, but this proposal is meant to do the same.

Regards,
Marco